requirements.txt
tkinter
pillow
python-docx
PyPDF2
pytesseract
openai